fix: address pr suggestions
