+++ internal/driver/driver.go
@@ -455,51 +455,6 @@ func (d *Driver) UpdateDevice(deviceName string, protocols map[string]models.Pro
 	return nil
 }
 
-// updateExistingDevice compares a discovered device and a matching existing device, and updates the existing
-// device network address and port if necessary
-func (d *Driver) updateExistingDevice(device models.Device, discDev sdkModel.DiscoveredDevice) error {
-	shouldUpdate := false
-	if device.OperatingState == models.Down {
-		device.OperatingState = models.Up
-		shouldUpdate = true
-	}
-
-	existAddr := device.Protocols[OnvifProtocol][Address]
-	existPort := device.Protocols[OnvifProtocol][Port]
-	discAddr := discDev.Protocols[OnvifProtocol][Address]
-	discPort := discDev.Protocols[OnvifProtocol][Port]
-	if discAddr == "" ||
-		discPort == "" ||
-		existAddr != discAddr ||
-		existPort != discPort {
-		d.lc.Info("Existing device has been discovered with a different network address.",
-			"oldInfo", fmt.Sprintf("%+v", existAddr+":"+existPort),
-			"discoveredInfo", fmt.Sprintf("%+v", discAddr+":"+discPort))
-
-		device.Protocols[OnvifProtocol][Address] = discAddr
-		device.Protocols[OnvifProtocol][Port] = discPort
-
-		device.OperatingState = models.Up
-		shouldUpdate = true
-	}
-
-	if !shouldUpdate {
-		// if both methods of dicovery are used, this message will print for the every camera discovered by netscan
-		d.lc.Warn("Re-discovered existing device at the same network address, nothing to do")
-		return nil
-	}
-
-	err := sdk.RunningService().UpdateDevice(device)
-	if err != nil {
-		d.lc.Error("There was an error updating the network address for an existing device.",
-			"deviceName", device.Name,
-			"error", err)
-		return errors.NewCommonEdgeXWrapper(err)
-	}
-
-	return nil
-}
-
 // RemoveDevice is a callback function that is invoked
 // when a Device associated with this Device Service is removed
 func (d *Driver) RemoveDevice(deviceName string, protocols map[string]models.ProtocolProperties) error {
